requirements.txt
gradio
torch
clip-openai